added testserver and commands i packagejson

+++ prisma/reset.ts
@@ -1,6 +1,9 @@
 import { db } from "./db";
 
 async function main() {
+	// NEVER ALLOW THIS OUTSIDE THE TEST ENVIRONMENT!!!
+	if (process.env.NODE_ENV !== "test") return;
+
 	await db.post.deleteMany({}); //lägg in för alla delarna, ta bort från alla tabeller så lägg till fler för varje här //
 }
 main()
